Rewriting topic to be task-focused
Changing title and intro to reflect the task the user is trying to
complete -- transferring or releasing a license, rather than the
procedure they have to follow -- unregistering the license.

Note the the procedure itself needs to change as it does not match the
interface. See https://github.com/oxygenxml/userguide/issues/8