Replacing literal installer names with keys
Replacing the literal installer names with keys.

In many cases, this enabled me to cut out a bunch of conditional logic
in the text.

However, in some cases, I had to remove <filepath> markup from around
the executable name in the text as <ph> is not legal inside <filepath>.

In one case this was not possible, however, as the name was part of a
complex file path:
\userguide\DITA\topics\segmentation-fault-macosx.dita. This is a problem
since it means we have a case of the install app as plain text which
might be missed if the name is changed and be believe that the key is
used in all cases.

Also note that there are other cases of oxygen archive and script files
that are referred to by name in plain text in the docs. For example,
see: C:\Users\Mark\userguide\DITA\topics\make-oxygen-portable.dita and
C:\Users\Mark\userguide\DITA\tasks\installation-instructions-all-platforms.dita
and
C:\Users\Mark\userguide\DITA\tasks\installation-instructions-linux-server.dita