Adding maps to preferences to avoid duplicate key defs
Some of the preferences topics appear in more than one place in the
preferences map to clarify the profiling of the preferences (one
profiled topic group per product). Some of these are the target of key
refs. By adding a map at the presentation topic level, we can define the
key once instead of repeating it.